move WidgetInfoProviderImpl class to internal package
